+++ test/pages/NetValueOfSalesFromNiPageSpec.scala
@@ -22,10 +22,10 @@ class NetValueOfSalesFromNiPageSpec extends PageBehaviours {
 
   "NetValueOfSalesFromNiPage" - {
 
-    beRetrievable[Int](NetValueOfSalesFromNiPage(index, index))
+    beRetrievable[BigDecimal](NetValueOfSalesFromNiPage(index, index))
 
-    beSettable[Int](NetValueOfSalesFromNiPage(index, index))
+    beSettable[BigDecimal](NetValueOfSalesFromNiPage(index, index))
 
-    beRemovable[Int](NetValueOfSalesFromNiPage(index, index))
+    beRemovable[BigDecimal](NetValueOfSalesFromNiPage(index, index))
   }
 }
